Redis presence manager
