+++ samples/ChatSample/Startup.cs
@@ -9,6 +9,7 @@ using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Identity.EntityFrameworkCore;
 using Microsoft.AspNetCore.SignalR;
+using Microsoft.AspNetCore.SignalR.Redis;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
@@ -54,15 +55,22 @@ namespace ChatSample
             services.AddTransient<IEmailSender, AuthMessageSender>();
             services.AddTransient<ISmsSender, AuthMessageSender>();
 
-            // To use Redis scaleout uncomment .AddRedis
+            // To use Redis scaleout uncomment .AddRedis and uncomment Redis related lines below for presence
             services.AddSignalR()
                 // .AddRedis()
                 ;
             services.AddAuthentication();
 
+
             services.AddSingleton(typeof(DefaultHubLifetimeManager<>), typeof(DefaultHubLifetimeManager<>));
             services.AddSingleton(typeof(HubLifetimeManager<>), typeof(DefaultPresenceHublifetimeMenager<>));
             services.AddSingleton(typeof(IUserTracker<>), typeof(InMemoryUserTracker<>));
+
+            /*
+            services.AddSingleton(typeof(RedisHubLifetimeManager<>), typeof(RedisHubLifetimeManager<>));
+            services.AddSingleton(typeof(HubLifetimeManager<>), typeof(RedisPresenceHublifetimeMenager<>));
+            services.AddSingleton(typeof(IUserTracker<>), typeof(RedisUserTracker<>));
+            */
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
